Faster Version of makeCAF for Quick Debugging
Note that this faster makeCAF does not use nusystematics to produce CAF variables; one advantage of it is that LArTPC and high pressure gas Ar TPC both have implemented detection threshold cuts and both follow a similar neutrino energy reconstruction scheme (event-by-event add up of all the visible energies); e.g. in this module, LArTPC does add in the mass of the pions towards reconstructing the neutrino energy as long as the pion is below LArTPC detection threshold.